streamlit_project/News_Summary/requirements.txt
streamlit
requests
transformers
torch
numpy
pandas
newspaper3k
lxml_html_clean